LPS-80123 blogs-uad - autogenerated
