+++ modules/apps/collaboration/blogs/blogs-uad/src/main/java/com/liferay/blogs/uad/exporter/BaseBlogsEntryUADExporter.java
@@ -1,0 +1,115 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.blogs.uad.exporter;
+
+import com.liferay.blogs.model.BlogsEntry;
+import com.liferay.blogs.service.BlogsEntryLocalService;
+import com.liferay.blogs.uad.constants.BlogsUADConstants;
+
+import com.liferay.portal.kernel.dao.orm.ActionableDynamicQuery;
+import com.liferay.portal.kernel.util.StringBundler;
+
+import com.liferay.user.associated.data.exporter.DynamicQueryUADExporter;
+
+import org.osgi.service.component.annotations.Reference;
+
+/**
+ * Provides the base implementation for the blogs entry UAD exporter.
+ *
+ * <p>
+ * This implementation exists only as a container for the default methods
+ * generated by ServiceBuilder. All custom service methods should be put in
+ * {@link BlogsEntryUADExporter}.
+ * </p>
+ *
+ * @author Brian Wing Shun Chan
+ * @generated
+ */
+public abstract class BaseBlogsEntryUADExporter extends DynamicQueryUADExporter<BlogsEntry> {
+	@Override
+	protected ActionableDynamicQuery doGetActionableDynamicQuery() {
+		return blogsEntryLocalService.getActionableDynamicQuery();
+	}
+
+	@Override
+	protected String[] doGetUserIdFieldNames() {
+		return BlogsUADConstants.USER_ID_FIELD_NAMES_BLOGS_ENTRY;
+	}
+
+	@Override
+	protected String toXmlString(BlogsEntry blogsEntry) {
+		StringBundler sb = new StringBundler(40);
+
+		sb.append("<model><model-name>");
+		sb.append("com.liferay.blogs.model.BlogsEntry");
+		sb.append("</model-name>");
+
+		sb.append(
+			"<column><column-name>entryId</column-name><column-value><![CDATA[");
+		sb.append(blogsEntry.getEntryId());
+		sb.append("]]></column-value></column>");
+		sb.append(
+			"<column><column-name>statusByUserId</column-name><column-value><![CDATA[");
+		sb.append(blogsEntry.getStatusByUserId());
+		sb.append("]]></column-value></column>");
+		sb.append(
+			"<column><column-name>statusByUserName</column-name><column-value><![CDATA[");
+		sb.append(blogsEntry.getStatusByUserName());
+		sb.append("]]></column-value></column>");
+		sb.append(
+			"<column><column-name>userId</column-name><column-value><![CDATA[");
+		sb.append(blogsEntry.getUserId());
+		sb.append("]]></column-value></column>");
+		sb.append(
+			"<column><column-name>userName</column-name><column-value><![CDATA[");
+		sb.append(blogsEntry.getUserName());
+		sb.append("]]></column-value></column>");
+		sb.append(
+			"<column><column-name>title</column-name><column-value><![CDATA[");
+		sb.append(blogsEntry.getTitle());
+		sb.append("]]></column-value></column>");
+		sb.append(
+			"<column><column-name>subtitle</column-name><column-value><![CDATA[");
+		sb.append(blogsEntry.getSubtitle());
+		sb.append("]]></column-value></column>");
+		sb.append(
+			"<column><column-name>urlTitle</column-name><column-value><![CDATA[");
+		sb.append(blogsEntry.getUrlTitle());
+		sb.append("]]></column-value></column>");
+		sb.append(
+			"<column><column-name>description</column-name><column-value><![CDATA[");
+		sb.append(blogsEntry.getDescription());
+		sb.append("]]></column-value></column>");
+		sb.append(
+			"<column><column-name>content</column-name><column-value><![CDATA[");
+		sb.append(blogsEntry.getContent());
+		sb.append("]]></column-value></column>");
+		sb.append(
+			"<column><column-name>smallImage</column-name><column-value><![CDATA[");
+		sb.append(blogsEntry.getSmallImage());
+		sb.append("]]></column-value></column>");
+		sb.append(
+			"<column><column-name>smallImageId</column-name><column-value><![CDATA[");
+		sb.append(blogsEntry.getSmallImageId());
+		sb.append("]]></column-value></column>");
+
+		sb.append("</model>");
+
+		return sb.toString();
+	}
+
+	@Reference
+	protected BlogsEntryLocalService blogsEntryLocalService;
+}
